refactor(test): extract shared golden-vector runner into support module
The four Phase-0 *_golden.rs suites (#2143-#2146) each re-implemented an
identical fixture loader, structural matcher, from() helper, and driver
loop. Extract them into core/tests/support/golden.rs, so each runner now
supplies only its util-specific run_case mapping plus a one-line test.

Behaviour and every existing golden case are unchanged.

Part of #2139
